Add files via upload
This is the SQL Server project for the original Hospital Management System. The purpose of this project is to create a database and corresponding tables as well as populate it with some sample data. This is done to mimic the scenario of having an existing database as I followed Database first approach for my original project HMS, stands for Hospital Management System